New handler registration pattern (#69)
* feat(core): update codegen and handler build process to use event emitter pattern for registering event handlers, wip

* feat: move from _app.ts to generated/index.ts

* chore(core): rename EventLog to Log

* feat(core): load schema before codegen

* chore(examples): update examples to new handler pattern

* fix: rm log

* feat(core): add warning logs for failed indexer queue tasks

* fix(core): make kill async

* test(core): back to directory based tests

* fix: typecheck failure when fixtures have not been generated

* feat(create-ponder): update to new handler pattern

* fix: malformed url

* test(core): jest config

* chore: changeset

* test(core): rm ponder dirs between tests and test for correct data inserted in basic example

+++ packages/core/src/db/entity/postgresEntityStore.ts
@@ -17,10 +17,11 @@ export class PostgresEntityStore implements EntityStore {
     this.db = db;
   }
 
-  async migrate(schema: Schema) {
+  async migrate(schema?: Schema) {
+    if (!schema) return;
     this.schema = schema;
 
-    schema.entities.forEach(async (entity) => {
+    this.schema.entities.forEach(async (entity) => {
       // Drop the table if it already exists
       await this.db.none(`DROP TABLE IF EXISTS "${entity.name}"`);
